Add support for dynamic xresources
This is a modified version of the xresources patch. Preferences are split into
two groups in `config.h`, `resources` and `scheme_resources`. The former are
loaded on startup from resources of the form

    st.shell: /bin/sh

The latter (by default all color settings) are loaded from resources of the form

    st.scheme.foo.color0: black

where `foo` is the name of the selected scheme, which can be set either on
invocation (`st -s foo`) or from a resource (`st.scheme: foo`). If not set
explicitly, the default scheme is called `default`.

Additionally, the scheme is exported as a window property named `ST_SCHEME`, and
modifying that property will dynamically reload the `scheme_resources`. This
makes it possible to switch the color scheme on runtime using sprop [1], either
from within the terminal,

    sprop $WINDOWID ST_SCHEME light

or from the outside by looping over all terminals,

    xdotool search --classname '^st-256color$' | xargs -I{} sprop {} ST_SCHEME light

[1]: https://tools.suckless.org/x/sprop/